refactor(solobase-native): extract load_dotenv / collect_app_env_vars / InfraConfig
Moves the generic env-handling into the solobase-native library; the
solobase-schema-specific known-keys filter moves to
solobase-server/src/config.rs as filter_to_declared_keys. The
library's collect_app_env_vars returns ALL non-SOLOBASE_-prefixed env
vars as a HashMap; the app applies filter_to_declared_keys before
seeding into the variables table.
